requirements.txt
fastapi
google-adk
supabase
python-dotenv
python-pptx
# fastapi[standard]